+++ @types/function/builtin.d.ts
@@ -1,5 +1,7 @@
 declare global {
   const builtin: {
     saveSettings: () => Promise<void>;
+    reloadEditor: (file: string, load_if_not_selected?: boolean) => void;
+    reloadEditorDebounced: (file: string, load_if_not_selected?: boolean) => void;
   };
 }