+++ src/function/builtin.ts
@@ -1,5 +1,8 @@
+import { reloadEditor, reloadEditorDebounced } from '@/compatibility';
 import { saveSettings } from '@sillytavern/script';
 
 export const builtin = {
   saveSettings,
+  reloadEditor,
+  reloadEditorDebounced,
 };